+++ examples/add-new-command/packages/sdk/src/doc/rpc/sample.ts
@@ -9,11 +9,13 @@
  *
  */
 
-export * from "./AbstractConfigManager";
-export * from "./ConfigFileUtils";
-export * from "./doc";
-export * from "./RpcClientApi";
-export * from "./SshConfigUtils";
-export * from "./SshErrors";
-export * from "./ZSshClient";
-export * from "./ZSshUtils";
+import type * as common from "./common";
+
+export interface PingRequest extends common.CommandRequest<"ping"> {
+    message?: string;
+}
+
+export interface PingResponse extends common.CommandResponse {
+    data: string;
+    timestamp: string;
+}