+++ native/zowed/server.hpp
@@ -18,6 +18,7 @@
 #include "../c/extend/plugin.hpp"
 #include "../c/singleton.hpp"
 #include "rpcio.hpp"
+#include "validator.hpp"
 
 // JSON-RPC 2.0 Standard Error Codes
 namespace RpcErrorCode
@@ -60,7 +61,7 @@ struct RpcResponse
   std::string jsonrpc;
   zstd::optional<zjson::Value> result;
   zstd::optional<ErrorDetails> error;
-  int id;
+  zstd::optional<int> id;
 };
 ZJSON_SERIALIZABLE(RpcResponse,
                    ZJSON_FIELD(RpcResponse, jsonrpc),
@@ -89,6 +90,8 @@ private:
   zjson::Value convert_output_to_json(const std::string &output);
   zjson::Value convert_ast_to_json(const ast::Node &ast_node);
   void print_response(const RpcResponse &response);
+  void print_error(int request_id, int code, const std::string &message, const std::string *data = nullptr);
+  validator::ValidationResult validate_json_with_schema(const std::string &method, const zjson::Value &params, bool is_request);
 
 public:
   /**